requirements.txt
pymongo
mysql-connector-python
python-snappy
